+++ server/models/index.js
@@ -3,5 +3,6 @@ const Location = require('./Location');
 const Instrument = require('./Instrument');
 const Filter = require('./Filter');
 const Session = require('./Session');
+const Category = require('./Category');
 
-module.exports = { User, Location, Instrument, Filter, Session };
+module.exports = { User, Location, Instrument, Filter, Session, Category };